rename module to messages
